internal/{device,sys}: suppress error logging for allowed missing devices
When a non-required device is marked used by a service and it is
missing, calls to the service will still result in attempts to get the
device to in order to complete an action. This will result in a logged
error. To prevent this, keep a note of all kernel services that depend
on allowed missing devices and return a sentinel error to ignore the
failure. This is only required for kernel services that interact with
input devices since daemon services do not have access to devices. Funcs
are updated to be aware of the error when they are device controllers.

Also remove an incorrectly added delete in sys.Manager.close.