cmd/watcher: fix handling of screen locking on more recent x11
Sometime between ubuntu 18.04 and 22.04 the behaviour of calls to
get_focused_window changed when the screen is locked. In the older
version the call succeeded when locked while it does not on the newer
version.

So move the call to get_focused_window lower in the C activeWindow
function to allow idle time to be recovered if get_focused_window does
not succeed, and do not return a warning error if we could not get
active window information and the screen was locked.