cmd/worklog/store: improve DB.Load behaviour
Make Load transactional and allow loading of new data into a store with existing
buckets with the same identity.


Also make CreateBuckets transactional, purely to indicate that returned bucket
metadata will be consistent with the backing database. All other compound store
queries are read-only and are protected by the DB mutex; all accesses to the
sqlite db are serialised as they are expected to be via DB in-process.