internal/{device,sys}: revise reconfiguration strategy
The previous design was fundamentally flawed, depending on rising edge state
changes which needed dubious hacks in order to be able to communicate the
removal of configured items. This changes the approach to a level state approach.
The original approach was intended to reduce churn during config editing, but
this was obviously misguided.

Change the behaviour of device.Button.Stop to repause after the image has been
blanked on stop so that ghost images are not rendered during configuration.