dex,internal/device: fix context cancellation for async action methods
The jsonrpc2.Handler cancels the context passed into to the Handle call on return
from the call. This means that asynchronous actions like drawing an animation on
a button or running a set of button updates from an RPC call will be terminated
despite being intended to run beyond the length of the call. We still need to let
the main context cancel actions in these cases, so make a second set of functions
that listen to the main context but ignore the handler context.