mod: update for go1.24
This fixes vet complaint about the use of non-constant format string in
types.NewErr calls, and go-cmp changes due to the movement of the internal
state of sync.Mutex into an internal package.

Script tests' regular expressions are updated to reflect the change in
build info version printing.

The watcher tester is moved into its own module to allow gioui.org to be
updated to handle go1.24 Cgo type method constraints. This also has the
advantage that gioui.org is now no longer a dependency of dex itself.